server/pyproject.toml
[project]
name = "server"
version = "1.0.0"
description = "Doctor Octopus Test Runner Application"
authors = [ { name = "Imran Sahil",  email = "imran.sahil@octaura.com" } ]
readme = "readme.md"
requires-python = ">=3.9"

dependencies = [
  "aiofiles==24.1.0",
  "annotated-types==0.7.0",
  "anyio==4.9.0",
  "bidict==0.23.1",
  "boto3==1.38.46",
  "botocore==1.38.46",
  "click==8.1.8",
  "fastapi-slim==0.115.14",
  "h11==0.16.0",
  "idna==3.10",
  "jmespath==1.0.1",
  "pydantic==2.11.7",
  "pydantic_core==2.33.2",
  "python-dateutil==2.9.0.post0",
  "python-dotenv==1.0.1",
  "python-engineio==4.12.2",
  "python-socketio==5.13.0",
  "redis==5.2.1",
  "s3transfer==0.13.0",
  "simple-websocket==1.1.0",
  "six==1.17.0",
  "sniffio==1.3.1",
  "starlette==0.46.2",
  "typing_extensions==4.14.0",
  "urllib3<=2.5.0",
  "uvicorn==0.35.0",
  "wsproto==1.2.0",
  "ruff==0.9.6"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
initialize = "initialize:main"
server = "server:fastapi_app"
server-fixme = "server-fixme:fastapi_app"

[tool.ruff]
line-length = 120
indent-width = 4

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]

# Ignore the follwing lint rules
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# Respect magic trailing commas.
skip-magic-trailing-comma = false

# Automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
# This is currently disabled by default, but it is planned for this to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in docstrings.
# This only has an effect when the `docstring-code-format` setting is enabled.
docstring-code-line-length = "dynamic"
